# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Farming Service API",
        "description": "API documentation for Farming Service",
        "version": "1.0.0"
    },
    "paths": {
        "/api/areas": {
            "get": {
                "tags": [
                    "Areas"
                ],
                "summary": "Get all areas",
                "operationId": "f16518e18a005a2f966b63c0e1745342",
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Areas"
                ],
                "summary": "Add a new area",
                "description": "Create a new area with city, state, village, pincode, and enabled status.",
                "operationId": "beeb06a4f56c4e31c8448c405163bed9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "city_id",
                                    "state_id",
                                    "pincode"
                                ],
                                "properties": {
                                    "city_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "state_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "village": {
                                        "type": "string",
                                        "example": "Greenfield"
                                    },
                                    "pincode": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "is_enabled": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/201"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/areas/{id}": {
            "get": {
                "tags": [
                    "Areas"
                ],
                "summary": "Get a specific area",
                "operationId": "d45f3822a5bdba2f44520ac32857bfb4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Area ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Areas"
                ],
                "summary": "Update an existing area",
                "operationId": "cc20a1af054acc1ea386968c8d9ba095",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Area ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "pincode": {
                                        "type": "string",
                                        "example": "654321"
                                    },
                                    "village": {
                                        "type": "string",
                                        "example": "Updated Village"
                                    },
                                    "is_enabled": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Areas"
                ],
                "summary": "Soft delete an area",
                "operationId": "820be3b502aa09c30f43b877c1c4527e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Area ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "responses": {
...
            "500": {
                "description": "Internal Server Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Something went wrong!"
                                },
                                "error": {
                                    "type": "string",
                                    "example": "Internal Server Error"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "200": {
                "description": "Successful operation",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": true
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Operation successful"
                                },
                                "data": {
                                    "type": "object",
                                    "example": []
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "201": {
                "description": "Resource created successfully",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": true
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Resource created successfully"
                                },
                                "data": {
                                    "type": "object",
                                    "example": []
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "422": {
                "description": "Validation error - invalid input data",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "error": {
                                    "type": "string",
                                    "example": "The given data was invalid."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}